# fulkoping_rental/src/main/java/se/gritacademy/fulkoping_rental/dto/user/UserDTO.java
package se.gritacademy.fulkoping_rental.dto.user;

public class UserDTO {
    private Long id;
    private String type;
    private String firstName;
    private String lastName;
    private String email;
    // Only Customer
    private String phoneNumber;
    // Only Admin
    private String employeeNumber;

    public UserDTO() {}

    public UserDTO(Long id, String type, String firstName, String lastName, String email,
                   String phoneNumber, String employeeNumber) {
        this.id = id;
        this.type = type;
        this.firstName = firstName;
        this.lastName = lastName;
        this.email = email;
        this.phoneNumber = phoneNumber;
        this.employeeNumber = employeeNumber;
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getType() {
        return type;
    }

    public void setType(String type) {
        this.type = type;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPhoneNumber() {
        return phoneNumber;
    }

    public void setPhoneNumber(String phoneNumber) {
        this.phoneNumber = phoneNumber;
    }

    public String getEmployeeNumber() {
        return employeeNumber;
    }

    public void setEmployeeNumber(String employeeNumber) {
        this.employeeNumber = employeeNumber;
    }
}